+++ platform/swift/source/CoreLogger.swift
@@ -174,8 +174,8 @@ extension CoreLogger: CoreLogging {
         self.underlyingLogger.notifyLowMemory(level: level, memoryUsedKB: memoryUsedKB, timestampUs: timestampUs)
     }
 
-    func registerOpaqueUserID(_ opaqueUserID: String) {
-        self.underlyingLogger.registerOpaqueUserID(opaqueUserID)
+    func registerOpaqueEntityID(_ opaqueEntityID: String) {
+        self.underlyingLogger.registerOpaqueEntityID(opaqueEntityID)
     }
 
     func runtimeValue<T: RuntimeValue>(_ variable: RuntimeVariable<T>) -> T {