+++ platform/swift/source/LoggerBridge.swift
@@ -260,8 +260,8 @@ final class LoggerBridge: LoggerBridging {
         capture_notify_low_memory(self.loggerID, level, memoryUsedKB, timestampUs)
     }
 
-    func registerOpaqueUserID(_ opaqueUserID: String) {
-        capture_register_opaque_user_id(self.loggerID, opaqueUserID)
+    func registerOpaqueEntityID(_ opaqueEntityID: String) {
+        capture_register_opaque_entity_id(self.loggerID, opaqueEntityID)
     }
 
     func runtimeValue<T: RuntimeValue>(_ variable: RuntimeVariable<T>) -> T {