+++ platform/swift/source/reports/BitdriftKSCrashWrapper.m
@@ -23,10 +23,17 @@
 #endif
 }
 
-+ (NSDictionary<NSString *, id> *)enhancedMetricKitReport:(NSDictionary<NSString *, id> *)metricKitReport {
++ (NSDictionary<NSString *, id> *)enhancedMetricKitReport:(NSDictionary<NSString *, id> *)metricKitReport
+                                      useStackOverlapMatching:(BOOL)useStackOverlapMatching
+                                                   summaryOut:(NSDictionary<NSString *, NSString *> * _Nullable * _Nullable)summaryOut {
 #ifndef BITDRIFT_OMIT_KSCRASH
-    return [BitdriftKSCrashHandler enhancedMetricKitReport:metricKitReport];
+    return [BitdriftKSCrashHandler enhancedMetricKitReport:metricKitReport
+                                  useStackOverlapMatching:useStackOverlapMatching
+                                               summaryOut:summaryOut];
 #else
+    if (summaryOut != nil) {
+        *summaryOut = nil;
+    }
     return metricKitReport;
 #endif
 }